refactoring for proper injection
