+++ src/remote/external.controller.service.interface.ts
@@ -1,0 +1,5 @@
+import { ExternalId } from '../db/entity/external.id';
+
+export abstract class IExternalControllerService {
+    abstract callExternalWalletCreate(externalIds: ExternalId[]): Promise<string>;
+}
