+++ test/e2e/escrow.fingerprint.e2e-spec.ts
@@ -16,6 +16,8 @@ import { MockRepository } from '../mock/mock.repository';
 import { ExternalId } from '../../src/db/entity/external.id';
 import { ExternalIdDbGateway } from '../../src/db/external.id.db.gateway';
 import { WalletCredentialsDbGateway } from '../../src/db/wallet.credentials.db.gateway';
+import { IExternalControllerService } from '../../src/remote/external.controller.service.interface';
+import { MockExternalControllerService } from '../mock/mock.external.controller.service';
 
 /**
  * This mocks out external dependencies (e.g. Db)
@@ -86,6 +88,10 @@ describe('EscrowController (e2e) using fingerprint plugin', () => {
                 {
                     provide: IBioAuthService,
                     useValue: mockBioAuthService
+                },
+                {
+                    provide: IExternalControllerService,
+                    useValue: MockExternalControllerService
                 },
                 PluginFactory,
             ],
